Switch terminal emulator to alacritty
For some reason, xfce4-terminal instantly exits when kakoune creates a
new window. Fiddling with it to try to get debug info yielded nothing,
except that running it without a daemon made it work most of the time.
Fiddling with it more caused it to crash, killing all my terminals.